Split Platform.cs into partial classes
